Changed format so that the frontend works for actual data (predictio data not implemented yet)
